backend/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
python-dotenv
passlib==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
google-auth
requests
stripe
python-multipart

pytest
httpx
pydantic[email]
yfinance
psycopg
pandas
pytickersymbols